rename/#50 auth usecase 일부 파일 이름 변경 및 경로 수정
